requirements.txt
absl-py
apache-beam
astunparse
attrs
avro-python3
backcall
cachetools
certifi
cffi
chardet
contextlib2
crcmod
cycler
dataclasses
decorator
dill
dm-tree
docopt
fastavro
future
gast
gin-config
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-cloud-bigquery
google-cloud-core
google-crc32c
google-pasta
google-resumable-media
googleapis-common-protos
grpcio
h5py
hdfs
httplib2
idna
imageio
ipython
ipython-genutils
jedi
joblib
kaggle
Keras
Keras-Preprocessing
keras-resnet
kiwisolver
lxml
Markdown
matplotlib
mock
numpy
oauth2client
oauthlib
object-detection
opencv-python
opencv-python-headless
opt-einsum
pandas
parso
pbr
pexpect
pickleshare
Pillow
progressbar2
promise
prompt-toolkit
protobuf
psutil
ptyprocess
py-cpuinfo
pyarrow
pyasn1
pyasn1-modules
pycocotools
pycparser
pydot
Pygments
pymongo
pyparsing
PyQt5
PyQt5-sip
python-dateutil
python-slugify
python-utils
pytz
PyYAML
requests
requests-oauthlib
rsa
scikit-learn
scipy
sentencepiece
six
slugify
tensorboard
tensorboard-plugin-wit
tensorflow
tensorflow-addons
tensorflow-datasets
tensorflow-estimator
tensorflow-hub
tensorflow-metadata
tensorflow-model-optimization
termcolor
text-unidecode
tf-models-official
tf-slim
threadpoolctl
tqdm
traitlets
typeguard
typing-extensions
uritemplate
urllib3
wcwidth
Werkzeug
wrapt
yolk
pynvml
tdigest
jsonpickle
flask-cors
tensorflowjs
attrdict
geventhttpclient